The `GameExhibit` object is actually used
Using this decorator lets us get rid of a smelly instance variable in `GamesController#show` (`@opponent`). which is rather nice.

The `games/show.html.erb` view template is now a rather empty shell for the `@game`'s partial, but this could change with the design (shared UI elements could be moved there, for example).